removing warnings - using correct types

+++ ios/ApplifierImpact/ApplifierImpactData/ApplifierImpactCache/ApplifierImpactCacheOperation.h
@@ -28,7 +28,7 @@ typedef enum {
 @interface ApplifierImpactCacheOperation : NSOperation
 
 @property (nonatomic, assign) id <ApplifierImpactCacheOperationDelegate> delegate;
-@property (nonatomic, assign) NSUInteger expectedFileSize;
+@property (nonatomic, assign) long long expectedFileSize;
 @property (nonatomic, copy)   NSString * operationKey;
 @property (nonatomic, assign) ResourceType resourceType;
